WIP: Use Azure.Messaging.EventHubs
This change moves the implementaion of the extension over to using the
new Azure.Messaging.EventHubs package, which is the newest SDK for Event
Hubs.

The design between Microsoft.Azure.Messaging.EventHubs (which was
previously used) and Azure.Messaging.EventHubs is very different when it
comes to actuall using a processor to read events and much of the design
of the function binding was built around the architecture of the older
SDK.

In order to bridge these gaps, I have introduced a custom processor,
trying to keep much of the shape of the older SDK in cases where the
design would not be observable to end users of the binding (for
example, we now have an internal EventProcessorHost type which builds
upon the new SDK to expose a processor API that feels similar to what
the older SDK has). In cases where this would cause an implementation
detail to leak out, I've biased towards following the design of the
newer SDK.

One interesting note here is around how checkpointing and ownership are
handled. The checkpointing and ownership strategy changed between the
two versions of the Event Hub SDKs. This binding writes ownership
and checkpoint information using the newer format, but supports reading
and writing checkpoints in the "legacy" format as well. When checkpoint
information exists in both forms, the newer checkpoint is used instead
of the old. This will allow migration from libraries written against the
older version of the trigger to the newer version of the trigger without
having to reprocess events while at the same time allowing systems (like
the Scale Controller) to see the current checkpoint in the older format
as they expect.